Fix missing ch6 subsection and TOC entry
I think there was an issue with the pandoc app converting html to LaTex for chapter 6.
Specifically with the missing subsection tag. This in turn caused the TOC to get generated
without chapter 6 referenced. This has been corrected as well as some cleanup to the
TOC file (delete  subsubsections) as well as taking the center tag out of pointers.htm